health check/requirements.txt
flask==3.1.0
gunicorn==23.0.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-containerregistry==1.2.0
pyodbc==5.2.0
redis==5.2.0